# docs/swagger.yaml
basePath: /
definitions:
...
  api.libraryFlow:
    properties:
      blockCount:
        type: integer
      description:
        type: string
      id:
        type: string
      isSharedWithMe:
        type: boolean
      name:
        type: string
      ownerDisplayName:
        type: string
      ownerId:
        type: string
      subflowCount:
        type: integer
      updatedAt:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is the API for the Pad Analyzer project.
  title: Pad Analyzer API
  version: "1.0"
paths:
...
  /api/library:
    get:
      description: Returns a list of flow documents stored in the library, with optional
        filtering by organization and search query.
      parameters:
      - description: Organization ID
        in: query
        name: orgId
        type: string
      - description: Search query
        in: query
        name: q
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.libraryFlow'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List library flows
      tags:
      - library
    post:
      consumes:
      - application/json
      description: Saves a new flow document to the library.
      parameters:
      - description: Create Library Flow Request
        in: body
        name: request
        required: true
        schema:
          properties:
            content:
              items:
                format: int32
                type: integer
              type: array
            description:
              type: string
            name:
              type: string
            orgId:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.libraryFlow'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create library flow
      tags:
      - library
  /api/library/{id}:
    delete:
      description: Deletes a specific flow document from the library. Only the owner
        can delete.
      parameters:
      - description: Flow ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete library flow
      tags:
      - library
    get:
      description: Returns the metadata for a specific library flow.
      parameters:
      - description: Flow ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.libraryFlow'
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get library flow metadata
      tags:
      - library
    put:
      consumes:
      - application/json
      description: Updates the name, description, or content of a specific library
        flow. Only the owner can update.
      parameters:
      - description: Flow ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Library Flow Request
        in: body
        name: request
        required: true
        schema:
          properties:
            content:
              items:
                format: int32
                type: integer
              type: array
            description:
              type: string
            name:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.libraryFlow'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update library flow
      tags:
      - library
  /api/library/{id}/content:
    get:
      description: Returns the raw JSON content of a specific library flow.
      parameters:
      - description: Flow ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get library flow content
      tags:
      - library
...
swagger: "2.0"